Code Review: Refactor/Quality
